pyproject.toml
[project]
name = "agentcon-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "helpers>=0.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "sentence-transformers>=5.2.0",
    "tiktoken>=0.12.0",
    "together>=1.5.31",
]
